+++ src/store/reducers/tenants/tenants.ts
@@ -1,26 +1,8 @@
-import {createSlice} from '@reduxjs/toolkit';
-import type {PayloadAction} from '@reduxjs/toolkit';
-
 import type {TTenantInfo} from '../../../types/api/tenant';
 import {api} from '../api';
 
-import type {PreparedTenant, TenantsState} from './types';
+import type {PreparedTenant} from './types';
 import {prepareTenants} from './utils';
-
-const initialState: TenantsState = {searchValue: ''};
-
-const slice = createSlice({
-    name: 'tenants',
-    initialState,
-    reducers: {
-        setSearchValue: (state, action: PayloadAction<string>) => {
-            state.searchValue = action.payload;
-        },
-    },
-});
-
-export const {setSearchValue} = slice.actions;
-export default slice.reducer;
 
 export const tenantsApi = api.injectEndpoints({
     endpoints: (build) => ({