+++ src/utils/monaco/constats.ts
@@ -1,2 +1,2 @@
-export const S_EXPRESSION_LANGUAGE_ID = 's-expression';
-export const YQL_LANGUAGE_ID = 'yql';
+export const S_EXPRESSION_LANGUAGE_ID = 's-expression_unstable';
+export const YQL_LANGUAGE_ID = 'yql_unstable';